requirements.txt
# Backend
django
djangorestframework
django-cors-headers
coreapi
python-dotenv

